Small changes from 1-20; Readme (#992)
* Update chap_00012.txt

Typo: "Are you doing to take coins from me?" to "Are you going to take coins from me?"

* Update chap_00006.txt

Typo: missing space. Image name left unchanged

* Update chap_00013.txt

Typo: head to heads

* Update chap_00017.txt

Punctuation

* Update README.md

Minor grammar adjustments and a few wording changes to tidy things up.

* Update chap_00019.txt

Changed: Kwajijijik (과지지직?) is probably an electricity or crackling sound effect. Susuk (스슥) can be the sound of movement or something passing by depending on context, and I interpreted it as the gesture of Mr Goblin swiftly obliterating the food out of existence. Not an ideal translation as I'm not Korean and I don't have the raws to reference, but less confusing than before.

* Update chap_00020.txt

* Update README.md